Changed Section class to SectionTool in utils Refactoring in CoursePlanner
